Experiments with quoting in ctest->bat->ps1
As of this commit,

    mkdir build-win
    cd build-win
    cmake ..
    ctest -V .

works in the sense that the entire command line, with commas, is stuffed
into `$MyInvocation.UnboundArugments[0]`.  This is the case whether
running ctest from cmd or PowerShell (console or ISE).  Example:

    PS > ctest . -V
    ... <lots cut>
        Start 1

    1: Test command: C:\Users\chrisw\proj-local\editorconfig-core-vimscript\pstest.bat "-f" "braces.in" "C:/Users/chrisw/proj-local/editorconfig-core-vimscript/tests/glob/a,.d"
    ... <lots cut>
    1: 0: >>-f braces.in C:/Users/chrisw/proj-local/editorconfig-core-vimscript/tests/glob/a,.d<<

The next step is to parse the command-line string according to POSIX
rules.  Or maybe switch to VBScript!